Addressed CircleCI and code review feedback. Also cleaned up some style issues.
