+++ server/status.go
@@ -626,12 +626,8 @@ func respondAsJSON(w http.ResponseWriter, r *http.Request, response interface{})
 	}
 }
 
-// Returns the path needed to issue a GET request for node status.
+// PathForNodeStatus returns the path needed to issue a GET request for node status. If passed
+// an empty nodeID, this returns the path to GET status for all nodes.
 func PathForNodeStatus(nodeID string) string {
 	return fmt.Sprintf("%s/%s", statusNodesPrefix, nodeID)
-}
-
-// Returns the path needed to issue a GET request for node status.
-func PathForAllNodesStatus() string {
-	return statusNodesPrefix
 }
