+++ util/http.go
@@ -18,6 +18,8 @@ package util
 
 import (
 	"encoding/json"
+	"fmt"
+	"io/ioutil"
 	"math"
 	"net/http"
 	"reflect"
@@ -238,4 +240,23 @@ func MarshalResponse(r *http.Request, value interface{}, allowed []EncodingType)
 		}
 	}
 	return
+}
+
+// GetJSON retrieves the URL specified by the parameters and and unmarshals the result into the
+// supplied interface.
+func GetJSON(httpClient *http.Client, scheme, hostport, path string, v interface{}) error {
+	url := fmt.Sprintf("%s://%s%s", scheme, hostport, path)
+	resp, err := httpClient.Get(url)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+	b, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return err
+	}
+	if err := json.Unmarshal(b, v); err != nil {
+		return err
+	}
+	return nil
 }
